feature: do components: popups

+++ src/utils/api.js
@@ -67,27 +67,27 @@ class Api {
         })
     }
 
-    likeCard(cardId) {
-        return fetch(`${this._address}/cards/${cardId}/likes`, {
-            method: 'PUT',
-            headers: {
-                authorization: this._token,
-                'Content-Type': 'application/json'
-            }
-        })
-            .then(this._checkResponse)
+    changeLikeCardStatus(cardId, isLiked) {
+        if (isLiked) {
+            return fetch(`${this._address}/cards/${cardId}/likes`, {
+                method: 'DELETE',
+                headers: {
+                    authorization: this._token
+                }
+            })
+                .then(this._checkResponse)
+        } else {
+            return fetch(`${this._address}/cards/${cardId}/likes`, {
+                method: 'PUT',
+                headers: {
+                    authorization: this._token,
+                    'Content-Type': 'application/json'
+                }
+            })
+                .then(this._checkResponse)
+        }
     }
-
-    deleteLikeCard(cardId) {
-        return fetch(`${this._address}/cards/${cardId}/likes`, {
-            method: 'DELETE',
-            headers: {
-                authorization: this._token
-            }
-        })
-            .then(this._checkResponse)
-    }
-
+    
     updateAvatar(link) {
         return fetch(`${this._address}/users/me/avatar`, {
             method: 'PATCH',
